+++ src/ForWebforms/ModelBinding/IBindingStrategy.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+
+namespace ExitStrategy.ForWebforms.ModelBinding
+{
+    public interface IBindingStrategy
+    {
+        ModelDefinition ExtractModel(IEnumerable dataSource);
+    }
+}
